app/requirements1.txt
fastapi
uvicorn[standard]
pydantic
alembic
sqlalchemy==1.4.32
psycopg2
pyJWT
pydantic[email]